@@ -1,28 +1,31 @@
 name: irri
 description: Third party, modern support for the Hydrawise API
-version: 0.0.23+26
+version: 0.0.24+27
 publish_to: none
 
 environment:
-  sdk: ">=2.13.0 <3.0.0"
+  sdk: ">=2.16.0 <3.0.0"
+  flutter: ">=2.10.2"
 
 dependencies:
-  bloc: ^7.1.0
+  bloc: ^8.0.3
+  clock: ^1.1.0
   cloud_firestore: ^3.1.0
   collection: ^1.15.0
   dio: ^4.0.0
-  firebase_analytics: ^8.3.4
-  firebase_auth: ^3.2.0
-  firebase_core: ^1.8.0
+  firebase_analytics: ^9.1.2
+  firebase_auth: ^3.3.9
+  firebase_core: ^1.13.1
   firebase_crashlytics: ^2.2.4
-  firebase_messaging: ^11.1.0
+  firebase_messaging: ^11.2.8
   flutter:
     sdk: flutter
-  flutter_bloc: ^7.2.0
+  flutter_bloc: ^8.0.1
   flutter_localizations:
     sdk: flutter
-  freezed_annotation: ^0.14.2
-  go_router: ^2.1.2
+  flutter_native_timezone: ^2.0.0
+  freezed_annotation: ^1.1.0
+  go_router: ^3.0.4
   google_fonts: ^2.1.0
   http: ^0.13.3
   intl: ^0.17.0
@@ -31,21 +34,21 @@ dependencies:
   provider: ^6.0.0
   result_type: ^0.1.0
   shared_preferences: ^2.0.7
-  sqflite: ^2.0.0+4
   uuid: ^3.0.5
   weather: ^2.0.1
 
 dev_dependencies:
-  bloc_test: ^8.1.0
+  bloc_test: ^9.0.3
   build_runner: ^2.1.2
+  charlatan: ^0.1.0
   fake_cloud_firestore: ^1.2.0
   firebase_auth_mocks: ^0.8.1
   flutter_lints: ^1.0.4
   flutter_test:
     sdk: flutter
-  freezed: ^0.14.2
-  json_serializable: ^4.1.0
-  mocktail: ^0.2.0
+  freezed: ^1.1.0
+  json_serializable: ^6.1.4
+  mocktail: ^0.3.0
 
 flutter:
   uses-material-design: true
