@@ -1,14 +1,10 @@
 import 'package:irri/core/core.dart';
-import 'package:irri/core/networking/http_client.dart';
 import 'package:irri/features/auth/auth.dart';
 import 'package:irri/features/customer_details/customer_details.dart';
 import 'package:result_type/result_type.dart';
 
-typedef GetCustomerDetails = Future<UseCaseResult<CustomerDetails, String>>
-    Function();
-
-class GetCustomerDetailsFromHydrawise {
-  GetCustomerDetailsFromHydrawise({
+class GetCustomerDetails {
+  GetCustomerDetails({
     required HttpClient httpClient,
     required GetApiKey getApiKey,
     required CustomerDetailsRepository repository,
@@ -36,55 +32,13 @@ class GetCustomerDetailsFromHydrawise {
     if (response.isSuccess) {
       final customerDetails = CustomerDetails.fromJson(response.success!);
 
-      final customerIdentification = customerDetails.toCustomer(apiKey);
+      final customer = customerDetails.toCustomer(apiKey);
 
-      await _repository.insertCustomer(customerIdentification);
+      await _repository.insertCustomer(customer);
 
       return Success(customerDetails);
     }
 
     return Failure("Can't fetch customer details");
-  }
-}
-
-class GetFakeCustomerDetails {
-  GetFakeCustomerDetails({
-    required CustomerDetailsRepository repository,
-  }) : _repository = repository;
-
-  final CustomerDetailsRepository _repository;
-
-  Future<UseCaseResult<CustomerDetails, String>> call() async {
-    final customers = await _repository.getCustomers();
-
-    if (customers.isEmpty) {
-      // Insert a dummy customer
-      final fakeCustomer = Customer(
-        activeControllerId: 1234,
-        customerId: 5678,
-        apiKey: '1212',
-        lastStatusUpdate: 1631330889,
-      );
-      await _repository.insertCustomer(fakeCustomer);
-    }
-
-    // Query again for simplicity
-    final customer = await _repository.getCustomer();
-
-    return Success(
-      CustomerDetails(
-        activeControllerId: customer.activeControllerId,
-        customerId: customer.customerId,
-        controllers: [
-          Controller(
-            name: 'Fake Controller',
-            lastContact: 1631616496,
-            serialNumber: '123456789',
-            id: 1234,
-            status: 'All good!',
-          ),
-        ],
-      ),
-    );
   }
 }
