@@ -1,5 +1,6 @@
+// coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'controller.dart';
 
@@ -184,7 +185,7 @@ class _$_Controller implements _Controller {
       @JsonKey(name: 'status') required this.status});
 
   factory _$_Controller.fromJson(Map<String, dynamic> json) =>
-      _$_$_ControllerFromJson(json);
+      _$$_ControllerFromJson(json);
 
   @override
   @JsonKey(name: 'name')
@@ -241,7 +242,7 @@ class _$_Controller implements _Controller {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$_$_ControllerToJson(this);
+    return _$$_ControllerToJson(this);
   }
 }
 
