@@ -10,7 +10,7 @@ class Customer with _$Customer {
     @JsonKey(name: 'customer_id') required int customerId,
     @JsonKey(name: 'api_key') required String apiKey,
     @JsonKey(name: 'last_status_update') required int lastStatusUpdate,
-    @Default(0) @JsonKey(name: 'time_zone_offset') int timeZoneOffsetMillis,
+    @Default('America/New_York') @JsonKey(name: 'time_zone') String? timeZone,
   }) = _Customer;
 
   factory Customer.fromJson(
