@@ -1,5 +1,6 @@
+// coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'customer_details.dart';
 
@@ -153,7 +154,7 @@ class _$_CustomerDetails implements _CustomerDetails {
       @JsonKey(name: 'controllers') required this.controllers});
 
   factory _$_CustomerDetails.fromJson(Map<String, dynamic> json) =>
-      _$_$_CustomerDetailsFromJson(json);
+      _$$_CustomerDetailsFromJson(json);
 
   @override
   @JsonKey(name: 'controller_id')
@@ -199,7 +200,7 @@ class _$_CustomerDetails implements _CustomerDetails {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$_$_CustomerDetailsToJson(this);
+    return _$$_CustomerDetailsToJson(this);
   }
 }
 
