@@ -1,5 +1,6 @@
+// coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'customer_status.dart';
 
@@ -155,7 +156,7 @@ class _$_CustomerStatus implements _CustomerStatus {
       @JsonKey(name: 'relays') required this.zones});
 
   factory _$_CustomerStatus.fromJson(Map<String, dynamic> json) =>
-      _$_$_CustomerStatusFromJson(json);
+      _$$_CustomerStatusFromJson(json);
 
   @override
   @JsonKey(name: 'nextpoll')
@@ -204,7 +205,7 @@ class _$_CustomerStatus implements _CustomerStatus {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$_$_CustomerStatusToJson(this);
+    return _$$_CustomerStatusToJson(this);
   }
 }
 
