@@ -1,5 +1,6 @@
+// coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'zone.dart';
 
@@ -201,8 +202,7 @@ class _$_Zone implements _Zone {
       @JsonKey(name: 'time') required this.secondsUntilNextRun,
       @JsonKey(name: 'run') required this.lengthOfNextRunTimeOrTimeRemaining});
 
-  factory _$_Zone.fromJson(Map<String, dynamic> json) =>
-      _$_$_ZoneFromJson(json);
+  factory _$_Zone.fromJson(Map<String, dynamic> json) => _$$_ZoneFromJson(json);
 
   @override
   @JsonKey(name: 'relay_id')
@@ -270,7 +270,7 @@ class _$_Zone implements _Zone {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$_$_ZoneToJson(this);
+    return _$$_ZoneToJson(this);
   }
 }
 
