@@ -6,18 +6,16 @@ part of 'zone.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_Zone _$_$_ZoneFromJson(Map json) {
-  return _$_Zone(
-    id: json['relay_id'] as int,
-    physicalNumber: json['relay'] as int,
-    name: json['name'] as String,
-    nextTimeOfWaterFriendly: json['timestr'] as String,
-    secondsUntilNextRun: json['time'] as int,
-    lengthOfNextRunTimeOrTimeRemaining: json['run'] as int,
-  );
-}
+_$_Zone _$$_ZoneFromJson(Map json) => _$_Zone(
+      id: json['relay_id'] as int,
+      physicalNumber: json['relay'] as int,
+      name: json['name'] as String,
+      nextTimeOfWaterFriendly: json['timestr'] as String,
+      secondsUntilNextRun: json['time'] as int,
+      lengthOfNextRunTimeOrTimeRemaining: json['run'] as int,
+    );
 
-Map<String, dynamic> _$_$_ZoneToJson(_$_Zone instance) => <String, dynamic>{
+Map<String, dynamic> _$$_ZoneToJson(_$_Zone instance) => <String, dynamic>{
       'relay_id': instance.id,
       'relay': instance.physicalNumber,
       'name': instance.name,
