@@ -1,5 +1,6 @@
+// coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'program.dart';
 
@@ -160,7 +161,7 @@ class _$_Program implements _Program {
       @JsonKey(ignore: true) this.runs});
 
   factory _$_Program.fromJson(Map<String, dynamic> json) =>
-      _$_$_ProgramFromJson(json);
+      _$$_ProgramFromJson(json);
 
   @override
   final String id;
@@ -207,7 +208,7 @@ class _$_Program implements _Program {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$_$_ProgramToJson(this);
+    return _$$_ProgramToJson(this);
   }
 }
 
