@@ -6,14 +6,12 @@ part of 'run_zone_response.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_RunZoneResponse _$_$_RunZoneResponseFromJson(Map json) {
-  return _$_RunZoneResponse(
-    message: json['message'] as String,
-    typeOfMessage: json['message_type'] as String,
-  );
-}
+_$_RunZoneResponse _$$_RunZoneResponseFromJson(Map json) => _$_RunZoneResponse(
+      message: json['message'] as String,
+      typeOfMessage: json['message_type'] as String,
+    );
 
-Map<String, dynamic> _$_$_RunZoneResponseToJson(_$_RunZoneResponse instance) =>
+Map<String, dynamic> _$$_RunZoneResponseToJson(_$_RunZoneResponse instance) =>
     <String, dynamic>{
       'message': instance.message,
       'message_type': instance.typeOfMessage,
