@@ -2,8 +2,6 @@ import 'dart:async';
 import 'dart:developer';
 
 import 'package:bloc/bloc.dart';
-import 'package:firebase_core/firebase_core.dart';
-import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:flutter/widgets.dart';
 import 'package:irri/app/app_bloc_observer.dart';
 import 'package:irri/app/domain/app_domain_factory.dart';
@@ -13,33 +11,38 @@ import 'package:irri/core/core.dart';
 import 'package:irri/features/customer_details/customer_details.dart';
 
 Future<void> main() async {
-  Bloc.observer = AppBlocObserver();
   FlutterError.onError = (details) {
-    log(details.exceptionAsString(), stackTrace: details.stack);
+    log(details.exceptionAsString());
   };
 
   await runZonedGuarded(
     () async {
-      WidgetsFlutterBinding.ensureInitialized();
-      await Firebase.initializeApp();
-      await FirebaseCrashlytics.instance.setCrashlyticsCollectionEnabled(false);
+      await BlocOverrides.runZoned(
+        () async {
+          WidgetsFlutterBinding.ensureInitialized();
+          // await Firebase.initializeApp();
+          // await FirebaseCrashlytics.instance
+          //     .setCrashlyticsCollectionEnabled(false);
 
-      final dataStorage = InMemoryStorage();
-      final repository = InMemoryCustomerDetailsRepository();
-      final providers = DevelopmentDomainFactory.build(
-        dataStorage: dataStorage,
-        repository: repository,
-      );
+          final dataStorage = InMemoryStorage();
+          final repository = InMemoryCustomerDetailsRepository();
+          final providers = DevelopmentDomainFactory.build(
+            dataStorage: dataStorage,
+            repository: repository,
+          );
 
-      final router = await BuildAppRouter().call();
+          final router = await BuildAppRouter().call();
 
-      runApp(
-        IrriApp(
-          router: router,
-          providers: providers,
-        ),
+          runApp(
+            IrriApp(
+              router: router,
+              providers: providers,
+            ),
+          );
+        },
+        blocObserver: AppBlocObserver(),
       );
     },
-    (error, stackTrace) => log(error.toString(), stackTrace: stackTrace),
+    (error, stackTrace) => log(error.toString()),
   );
 }
