@@ -1,22 +1,31 @@
 import 'dart:async';
 
 import 'package:bloc_test/bloc_test.dart';
+import 'package:clock/clock.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:irri/core/core.dart';
 import 'package:irri/features/auth/auth.dart';
-import 'package:irri/features/auth/domain/validate_api_key.dart';
 import 'package:irri/features/customer_details/customer_details.dart';
 
 void main() {
   group('CustomerDetailsCubit', () {
     final DataStorage dataStorage = InMemoryStorage();
 
+    final repository = InMemoryCustomerDetailsRepository();
+    final setApiKey = SetApiKey(dataStorage);
+    final setFirebaseUid = SetFirebaseUid(dataStorage);
     late AuthCubit authCubit;
 
     CustomerDetailsCubit _buildSubject() {
-      final repository = InMemoryCustomerDetailsRepository();
-      final setApiKey = SetApiKey(dataStorage);
-      final setFirebaseUid = SetFirebaseUid(dataStorage);
+      return CustomerDetailsCubit(
+        getCustomerDetails: GetFakeCustomerDetails(repository: repository),
+        getCustomerStatus: GetFakeCustomerStatus(repository: repository),
+        authCubit: authCubit,
+      );
+    }
+
+    setUp(() async {
+      await dataStorage.clearAll();
 
       authCubit = AuthCubit(
         logOut: LogOut(
@@ -42,20 +51,6 @@ void main() {
           getFirebaseUid: GetFirebaseUid(dataStorage),
         ),
       );
-
-      return CustomerDetailsCubit(
-        getCustomerDetails: GetFakeCustomerDetails(
-          repository: repository,
-        ),
-        getCustomerStatus: GetFakeCustomerStatus(
-          repository: repository,
-        ),
-        authCubit: authCubit,
-      );
-    }
-
-    setUp(() async {
-      await dataStorage.clearAll();
     });
 
     group('start', () {
@@ -69,19 +64,27 @@ void main() {
       });
 
       group('when logged in', () {
+        // Uses a fixed wall-clock because customer status
+        // uses the current time for `timeOfLastStatusUnixEpoch`
+        final fixedWallClock = Clock.fixed(DateTime(2020));
         blocTest<CustomerDetailsCubit, CustomerDetailsState>(
           'it emits Complete',
           setUp: () async {
             await authCubit.login('fake-api-key');
           },
           build: _buildSubject,
-          act: (cubit) => cubit.start(),
+          act: (cubit) {
+            // Use fixed clock for the `act` zone
+            withClock(fixedWallClock, () {
+              cubit.start();
+            });
+          },
           expect: () => <CustomerDetailsState>[
             CustomerDetailsState.loading(),
             CustomerDetailsState.complete(
               customerDetails: CustomerDetails(
-                activeControllerId: 1234,
-                customerId: 5678,
+                activeControllerId: 1,
+                customerId: 1,
                 controllers: [
                   Controller(
                     name: 'Fake Controller',
@@ -94,7 +97,8 @@ void main() {
               ),
               customerStatus: CustomerStatus(
                 numberOfSecondsUntilNextRequest: 5,
-                timeOfLastStatusUnixEpoch: 1631330889,
+                timeOfLastStatusUnixEpoch:
+                    fixedWallClock.now().millisecondsSinceEpoch,
                 zones: [],
               ),
             ),
